requirements.txt
aiofiles==0.7.0
anyio==3.3.2
certifi==2021.5.30
charset-normalizer==2.0.6
flake8==3.9.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
mccabe==0.6.1
Pillow==8.3.2
pycodestyle==2.7.0
pyflakes==2.3.1
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
urllib3==1.26.7
